requirements.txt
aioredis==1.3.1
asgiref==3.5.0
astroid==2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.12.3
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.3
boto3==1.14.52
botocore==1.17.52
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.2
daphne==3.0.2
Django==3.2.11
django-cors-headers==3.4.0
django-storages==1.10
djangorestframework==3.11.2
docutils==0.15.2
googleapis-common-protos==1.52.0
hiredis==1.1.0
html-sanitizer==1.9.1
httplib2==0.19.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
mixpanel==4.7.0
msgpack==1.0.0
Pillow==9.0.0
protobuf==3.12.4
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.7.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2020.1
requests==2.24.0
rsa==4.7
s3transfer==0.3.3
sendgrid==6.6.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
stripe==2.55.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wrapt==1.12.1
zope.interface==5.4.0
gunicorn==20.1.0
